requirements.txt
discord.py==2.3.2
python-telegram-bot>=21.0
aiohttp==3.9.5
aiofiles==23.2.1